+++ packages/solid-styled/compiler/core/checks.ts
@@ -1,0 +1,43 @@
+import * as t from '@babel/types';
+
+export function getImportSpecifierKey(
+  specifier: t.ImportSpecifier,
+) {
+  if (t.isIdentifier(specifier.imported)) {
+    return specifier.imported.name;
+  }
+  return specifier.imported.value;
+}
+
+type TypeFilter<V extends t.Node> = (node: t.Node) => node is V;
+
+export function isPathValid<V extends t.Node>(
+  path: unknown,
+  key: TypeFilter<V>,
+): path is babel.NodePath<V> {
+  return key((path as babel.NodePath).node);
+}
+
+export type NestedExpression =
+  | t.ParenthesizedExpression
+  | t.TypeCastExpression
+  | t.TSAsExpression
+  | t.TSSatisfiesExpression
+  | t.TSNonNullExpression
+  | t.TSInstantiationExpression
+  | t.TSTypeAssertion;
+
+export function isNestedExpression(node: t.Node): node is NestedExpression {
+  switch (node.type) {
+    case 'ParenthesizedExpression':
+    case 'TypeCastExpression':
+    case 'TSAsExpression':
+    case 'TSSatisfiesExpression':
+    case 'TSNonNullExpression':
+    case 'TSTypeAssertion':
+    case 'TSInstantiationExpression':
+      return true;
+    default:
+      return false;
+  }
+}
