+++ packages/solid-styled/compiler/core/constants.ts
@@ -1,0 +1,15 @@
+// The identifier for the tagged template
+export const TAGGED_TEMPLATE = 'css';
+// The module
+export const SOURCE_MODULE = 'solid-styled';
+// The namespace used for scoping
+export const SOLID_STYLED_NS = 's';
+export const VARS_ID = 'vars';
+export const SHEET_ID = 'sheet';
+export const GLOBAL_SELECTOR = 'global';
+
+export const RUNTIME_IDENTIFIERS = {
+  createCSSVars: 'createCSSVars',
+  mergeStyles: 'mergeStyles',
+  useSolidStyled: 'useSolidStyled',
+};
