+++ packages/solid-styled/compiler/core/get-root-statement-path.ts
@@ -1,0 +1,13 @@
+import * as babel from '@babel/core';
+
+export default function getRootStatementPath(path: babel.NodePath) {
+  let current = path.parentPath;
+  while (current) {
+    const next = current.parentPath;
+    if (next && next.type === 'Program') {
+      return current;
+    }
+    current = next;
+  }
+  return path;
+}
