+++ packages/solid-styled/compiler/env.d.ts
