+++ packages/solid-styled/postcss.config.js
@@ -1,0 +1,1 @@
+export default {};
